+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # withings-cache
 
 ## Unreleased
+- task
+- meas
+
+
+## 0.3.3 - 2022-12-19
+fetch meas
+- checked git configurations
+- curl-get
+- curl-post
+- refresh-all-pmap!
+- fetch-weight
+- fetch-meas
+- fetch-meas-with
+- can fetch
+
+## 0.2.2 - 2022-12-19
+- login, users, tokens
 
 ## 0.1.0 - 2022-12-19
 - started melt repository